# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mongly API",
    "version": "0.6.0",
    "description": "2026-09-24 계약. 선택 50 grapheme 일기(본인·현재 친구 공개), KST 오늘 병 편집, 무제한 보관/페이지 조회, 친구 좋아요, FRIEND_REQUEST/JAR_LIKED 통합 알림. 오류는 {error:{code,message,details?}}. httpOnly mongly_token 쿠키 인증. 같은 출처 /api/*로 호출. 상태 변경 본문은 application/json. 기록당 구슬 1~7개·KST 날짜당 병 1개·친구 10명은 유지. 최신 상세 및 변경 안내는 docs/API.md, docs/API_CHANGES_2026-09-24.md."
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mongly_token",
        "description": "로그인/가입으로 발급되는 httpOnly 쿠키"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "JarEmotion": {
        "type": "object",
        "properties": {
          "emotionId": {
            "type": "integer",
            "minimum": 1
          },
          "name": {
            "type": "string"
          },
          "colorHex": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 7
          }
        },
        "required": [
          "emotionId",
          "name",
          "colorHex",
          "count"
        ]
      },
      "Jar": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "불투명 ID"
          },
          "recordDate": {
            "type": "string",
            "format": "date",
            "description": "서버 KST 날짜 YYYY-MM-DD"
          },
          "note": {
            "$ref": "#/components/schemas/Note"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "UTC ISO 8601"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "UTC ISO 8601"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "description": "실제 note/감정 변경마다 증가. 좋아요/읽음/no-op으로 변하지 않음"
          },
          "canEdit": {
            "type": "boolean",
            "description": "조회 시 본인이면서 KST 오늘 기록. 저장 시 날짜·권한 재검사"
          },
          "canLike": {
            "type": "boolean",
            "description": "현재 친구의 병이면 true, 본인 병은 false"
          },
          "likeCount": {
            "type": "integer",
            "minimum": 0
          },
          "likedByMe": {
            "type": "boolean"
          },
          "dominantEmotionId": {
            "type": "integer",
            "minimum": 1,
            "nullable": true,
            "description": "최다 감정. 동률이면 null"
          },
          "emotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JarEmotion"
            },
            "minItems": 1,
            "maxItems": 7,
            "description": "emotionId 고유, count 합계 1~7"
          }
        },
        "required": [
          "id",
          "recordDate",
          "note",
          "createdAt",
          "updatedAt",
          "version",
          "canEdit",
          "canLike",
          "likeCount",
          "likedByMe",
          "dominantEmotionId",
          "emotions"
        ],
        "description": "본인/친구 목록·상세·완료·수정 공통 응답. viewer에 따라 권한·likedByMe만 달라진다."
      },
      "Draft": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 7
          },
          "dominantEmotionId": {
            "type": "integer",
            "minimum": 1,
            "nullable": true
          },
          "emotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JarEmotion"
            },
            "maxItems": 7
          }
        },
        "required": [
          "total",
          "dominantEmotionId",
          "emotions"
        ]
      },
      "TodayState": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "jar": {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              },
              "draft": {
                "$ref": "#/components/schemas/Draft"
              }
            },
            "required": [
              "jar",
              "draft"
            ]
          },
          {
            "type": "object",
            "properties": {
              "jar": {
                "$ref": "#/components/schemas/Jar"
              },
              "draft": {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            },
            "required": [
              "jar",
              "draft"
            ]
          }
        ],
        "description": "완료 전 jar=null/draft=Draft, 완료 후 jar=Jar/draft=null"
      },
      "DraftEmotionInput": {
        "type": "object",
        "properties": {
          "emotionId": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "emotionId"
        ]
      },
      "CompleteJarInput": {
        "type": "object",
        "properties": {
          "note": {
            "$ref": "#/components/schemas/Note"
          }
        },
        "additionalProperties": false,
        "description": "본문 생략과 {}도 허용. 감정은 서버 드래프트 사용"
      },
...
      "EditJarInput": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "note": {
            "$ref": "#/components/schemas/Note"
          },
          "emotions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EditEmotionInput"
            },
            "minItems": 1,
            "maxItems": 7,
            "description": "전체 대체. emotionId 고유, count 합계 1~7. 기존 비활성 감정은 기존 count 이하만 허용"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "additionalProperties": false,
        "anyOf": [
          {
            "required": [
              "note"
            ]
          },
          {
            "required": [
              "emotions"
            ]
          }
        ],
        "description": "note 또는 emotions 하나 이상 필수. note 생략=유지, null=지우기. 버전 일치 확인 후 내용이 같으면 no-op"
      },
      "JarPage": {
        "type": "object",
        "properties": {
          "jars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Jar"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "다음 페이지가 없으면 null. 재사용 시 URL 인코딩하고 내용을 해석하지 않는다."
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "jars",
          "nextCursor",
          "hasMore"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/jars/today": {
      "get": {
        "tags": [
          "유리병"
        ],
        "summary": "서버 KST 오늘 상태",
        "responses": {
          "200": {
            "description": "완료 전 Draft 또는 완료 후 Jar. jar/draft 중 정확히 하나가 null.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayState"
                },
                "examples": {
                  "empty": {
                    "value": {
                      "jar": null,
                      "draft": {
                        "total": 0,
                        "dominantEmotionId": null,
                        "emotions": []
                      }
                    }
                  },
                  "draft": {
                    "value": {
                      "jar": null,
                      "draft": {
                        "total": 2,
                        "dominantEmotionId": 1,
                        "emotions": [
                          {
                            "emotionId": 1,
                            "name": "기쁨",
                            "colorHex": "#FFD54A",
                            "count": 2
                          }
                        ]
                      }
                    }
                  },
                  "completed": {
                    "value": {
                      "jar": {
                        "id": "cmjar0001",
                        "recordDate": "2026-09-24",
                        "note": "발표를 무사히 마쳐서 후련했다.",
                        "createdAt": "2026-09-24T10:00:00.000Z",
                        "updatedAt": "2026-09-24T10:00:00.000Z",
                        "version": 1,
                        "canEdit": true,
                        "canLike": false,
                        "likeCount": 0,
                        "likedByMe": false,
                        "dominantEmotionId": 1,
                        "emotions": [
                          {
                            "emotionId": 1,
                            "name": "기쁨",
                            "colorHex": "#FFD54A",
                            "count": 2
                          },
                          {
                            "emotionId": 6,
                            "name": "사랑",
                            "colorHex": "#FF78AE",
                            "count": 1
                          }
                        ]
                      },
                      "draft": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jars/draft/emotions": {
      "post": {
        "tags": [
          "유리병"
        ],
        "summary": "감정 구슬 1개 즉시 저장",
        "description": "프런트는 담기·되돌리기 요청을 직렬화한다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftEmotionInput"
              },
              "example": {
                "emotionId": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "갱신된 전체 드래프트",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                },
                "example": {
                  "total": 2,
                  "dominantEmotionId": 1,
                  "emotions": [
                    {
                      "emotionId": 1,
                      "name": "기쁨",
                      "colorHex": "#FFD54A",
                      "count": 2
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION / INVALID_EMOTION",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INVALID_EMOTION",
                    "message": "존재하지 않는 감정이에요."
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          },
          "409": {
            "description": "DRAFT_FULL / JAR_ALREADY_TODAY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "DRAFT_FULL",
                    "message": "감정은 최대 7개까지 담을 수 있어요."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "유리병"
        ],
        "summary": "마지막 구슬 1개 되돌리기",
        "description": "요청 본문 없음. Content-Type도 생략 가능.",
        "responses": {
          "200": {
            "description": "갱신된 전체 드래프트",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                },
                "example": {
                  "total": 0,
                  "dominantEmotionId": null,
                  "emotions": []
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          },
          "409": {
            "description": "되돌릴 구슬 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "DRAFT_EMPTY",
                    "message": "되돌릴 감정이 없어요."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jars": {
      "post": {
        "tags": [
          "유리병"
        ],
        "summary": "드래프트와 선택 일기로 오늘 병 완성",
        "description": "본문 생략/{} /note:null 허용. 감정은 서버 드래프트에서 읽는다. 기존 병은 개수와 무관하게 보존. 응답 유실 시 GET /jars/today로 확인하며 재호출은 기존 내용을 덮어쓰지 않는다.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteJarInput"
              },
              "example": {
                "note": "발표를 무사히 마쳐서 후련했다."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "병·일기 저장, 드래프트 정리",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Jar"
                },
                "example": {
                  "id": "cmjar0001",
                  "recordDate": "2026-09-24",
                  "note": "발표를 무사히 마쳐서 후련했다.",
                  "createdAt": "2026-09-24T10:00:00.000Z",
                  "updatedAt": "2026-09-24T10:00:00.000Z",
                  "version": 1,
                  "canEdit": true,
                  "canLike": false,
                  "likeCount": 0,
                  "likedByMe": false,
                  "dominantEmotionId": 1,
                  "emotions": [
                    {
                      "emotionId": 1,
                      "name": "기쁨",
                      "colorHex": "#FFD54A",
                      "count": 2
                    },
                    {
                      "emotionId": 6,
                      "name": "사랑",
                      "colorHex": "#FF78AE",
                      "count": 1
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION / DRAFT_EMPTY. 검증 실패 시 드래프트 보존",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "DRAFT_EMPTY",
                    "message": "담은 감정이 없어요."
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          },
          "409": {
            "description": "JAR_ALREADY_TODAY / 비정상 드래프트의 DRAFT_FULL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "JAR_ALREADY_TODAY",
                    "message": "오늘의 유리병은 이미 완성했어요."
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "유리병"
        ],
        "summary": "내 보관함 페이지 (저장 총량 무제한)",
        "description": "recordDate 내림차순. 기본 limit=7, 최대50. total은 제공하지 않는다. nextCursor=null이면 마지막. 커서 병이 삭제되어도 날짜 경계로 이동 가능. 최신 기록은 첫 페이지 재조회.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 7
            },
            "description": "정수 1~50. 저장 총량 한도가 아님"
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            },
            "description": "응답 nextCursor를 그대로 재전송. 버전·recordDate를 인코딩한 불투명 문자열"
          }
        ],
        "responses": {
          "200": {
            "description": "내 병 페이지",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JarPage"
                },
                "example": {
                  "jars": [
                    {
                      "id": "cmjar0001",
                      "recordDate": "2026-09-24",
                      "note": "발표를 무사히 마쳐서 후련했다.",
                      "createdAt": "2026-09-24T10:00:00.000Z",
                      "updatedAt": "2026-09-24T10:00:00.000Z",
                      "version": 1,
                      "canEdit": true,
                      "canLike": false,
                      "likeCount": 0,
                      "likedByMe": false,
                      "dominantEmotionId": 1,
                      "emotions": [
                        {
                          "emotionId": 1,
                          "name": "기쁨",
                          "colorHex": "#FFD54A",
                          "count": 2
                        },
                        {
                          "emotionId": 6,
                          "name": "사랑",
                          "colorHex": "#FF78AE",
                          "count": 1
                        }
                      ]
                    }
                  ],
                  "nextCursor": null,
                  "hasMore": false
                }
              }
            }
          },
          "400": {
            "description": "본문·쿼리 형식 오류 (VALIDATION)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "message": "요청 형식이 올바르지 않아요."
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jars/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "불투명 ID"
          },
          "description": "병 ID",
          "example": "cmjar0001"
        }
      ],
      "get": {
        "tags": [
          "유리병"
        ],
        "summary": "병 상세 (본인 또는 현재 친구)",
        "description": "친구의 과거 병도 ID로 열람 가능. note를 포함한 공통 Jar.",
        "responses": {
          "200": {
            "description": "병 상세",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Jar"
                },
                "example": {
                  "id": "cmjar0001",
                  "recordDate": "2026-09-24",
                  "note": "발표를 무사히 마쳐서 후련했다.",
                  "createdAt": "2026-09-24T10:00:00.000Z",
                  "updatedAt": "2026-09-24T10:00:00.000Z",
                  "version": 1,
                  "canEdit": true,
                  "canLike": false,
                  "likeCount": 0,
                  "likedByMe": false,
                  "dominantEmotionId": 1,
                  "emotions": [
                    {
                      "emotionId": 1,
                      "name": "기쁨",
                      "colorHex": "#FFD54A",
                      "count": 2
                    },
                    {
                      "emotionId": 6,
                      "name": "사랑",
                      "colorHex": "#FF78AE",
                      "count": 1
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "본문·쿼리 형식 오류 (VALIDATION)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "message": "요청 형식이 올바르지 않아요."
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          },
          "403": {
            "description": "본인·친구 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "친구의 유리병만 볼 수 있어요."
                  }
                }
              }
            }
          },
          "404": {
            "description": "병 없음·삭제됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "JAR_NOT_FOUND",
                    "message": "존재하지 않는 유리병이에요."
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "유리병"
        ],
        "summary": "작성자의 KST 오늘 병 감정·일기 편집",
        "description": "expectedVersion 필수, note/emotions 중 하나 이상. emotions는 전체 대체(고유 emotionId, count 합1~7). 잠금 후 서버 날짜·권한 재검사. 버전 불일치는 내용이 같아도409. 일치하고 정규화된 내용이 같으면 no-op으로 version/updatedAt 유지. 실제 수정 시 둘만 갱신하며 ID/recordDate/createdAt/좋아요/알림 보존. 기존 비활성 감정은 기존 수량 이하만 허용.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditJarInput"
              },
              "example": {
                "expectedVersion": 1,
                "note": "저녁에는 친구들과 함께 쉬었다.",
                "emotions": [
                  {
                    "emotionId": 1,
                    "count": 2
                  },
                  {
                    "emotionId": 6,
                    "count": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정본 또는 동일 내용 원본",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Jar"
                },
                "example": {
                  "id": "cmjar0001",
                  "recordDate": "2026-09-24",
                  "note": "저녁에는 친구들과 함께 쉬었다.",
                  "createdAt": "2026-09-24T10:00:00.000Z",
                  "updatedAt": "2026-09-24T10:15:00.000Z",
                  "version": 2,
                  "canEdit": true,
                  "canLike": false,
                  "likeCount": 0,
                  "likedByMe": false,
                  "dominantEmotionId": 1,
                  "emotions": [
                    {
                      "emotionId": 1,
                      "name": "기쁨",
                      "colorHex": "#FFD54A",
                      "count": 2
                    },
                    {
                      "emotionId": 6,
                      "name": "사랑",
                      "colorHex": "#FF78AE",
                      "count": 1
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION / INVALID_EMOTION",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "message": "요청 형식이 올바르지 않아요."
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          },
          "403": {
            "description": "작성자 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "내 유리병만 수정할 수 있어요."
                  }
                }
              }
            }
          },
          "404": {
            "description": "병 없음·삭제됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "JAR_NOT_FOUND",
                    "message": "존재하지 않는 유리병이에요."
                  }
                }
              }
            }
          },
          "409": {
            "description": "JAR_VERSION_CONFLICT(최신 상세 재조회) / JAR_EDIT_WINDOW_CLOSED(KST 자정 경과 포함)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "conflict": {
                    "value": {
                      "error": {
                        "code": "JAR_VERSION_CONFLICT",
                        "message": "다른 곳에서 기록이 수정됐어요. 최신 기록을 다시 확인해주세요.",
                        "details": {
                          "currentVersion": 2
                        }
                      }
                    }
                  },
                  "closed": {
                    "value": {
                      "error": {
                        "code": "JAR_EDIT_WINDOW_CLOSED",
                        "message": "오늘의 유리병만 수정할 수 있어요."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "유리병"
        ],
        "summary": "본인 병 삭제",
        "description": "요청 본문 없음. Content-Type도 생략 가능. 병의 좋아요·연결 알림도 삭제.",
        "responses": {
          "200": {
            "description": "삭제 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "description": "본문·쿼리 형식 오류 (VALIDATION)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION",
                    "message": "요청 형식이 올바르지 않아요."
                  }
                }
              }
            }
          },
          "401": {
            "description": "비로그인·만료·탈퇴 세션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "로그인이 필요해요."
                  }
                }
              }
            }
          },
          "403": {
            "description": "작성자 아님",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "내 유리병만 삭제할 수 있어요."
                  }
                }
              }
            }
          },
          "404": {
            "description": "병 없음·삭제됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "JAR_NOT_FOUND",
                    "message": "존재하지 않는 유리병이에요."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}